refactor: Move wallet subject subscription out of its model definition to break model/service dependency on controller
